+++ persistence-modules/spring-data-jpa-3/src/main/java/com/baeldung/model/Payment.java
@@ -1,4 +1,4 @@
-package com.baeldung.tx;
+package com.baeldung.model;
 
 import javax.persistence.*;
